OMOPCDMCohortCreator API

This is a list of documentation associated with every single exported function from OMOPCDMCohortCreator. There are a few different sections with a brief explanation of what these sections are followed by relevant functions.

OMOPCDMCohortCreator API

Getters

Filters

Generators

Executors

Welcome to the OMOPCDMCohortCreator.jl Docs!

Tutorials

Beginner Tutorial

Getters

This family of functions are dedicated to only getting information concerning a patient or OMOP CDM database.

OMOPCDMCohortCreator.GetDatabasePersonIDs — Function

GetDatabasePersonIDs(conn; tab::SQLTable = person)

Get all unique person_id's from a database.

Arguments:

conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Person table; default person

Returns

ids::Vector{Int64} - the list of persons

OMOPCDMCohortCreator.GetPatientState — Function

GetPatientState(ids, conn; tab::SQLTable = location, join_tab::SQLTable = person)

Given a list of person IDs, find their home state.

Arguments:

ids - list of person_id's; each ID must be of subtype Integer conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Location table; default location

join_tab::SQLTable - the SQLTable representing the Person table; default person

Returns

df::DataFrame - a two column DataFrame comprised of columns: :person_id and :state

OMOPCDMCohortCreator.GetPatientGender — Function

GetPatientGender(ids, conn; tab::SQLTable = person)

Given a list of person IDs, find their gender.

Arguments:

ids - list of person_id's; each ID must be of subtype Integer conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Person table; default person

Returns

df::DataFrame - a two column DataFrame comprised of columns: :person_id and :gender_concept_id

OMOPCDMCohortCreator.GetPatientRace — Function

GetPatientRace(ids, conn; tab::SQLTable = person)

Given a list of person IDs, find their race.

Arguments:

ids - list of person_id's; each ID must be of subtype Integer conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Person table; default person

Returns

df::DataFrame - a two column DataFrame comprised of columns: :person_id and :race_concept_id

OMOPCDMCohortCreator.GetPatientAgeGroup — Function

GetPatientAgeGroup( ids, conn; minuend = :now, age_groupings = [ [0, 9], [10, 19], [20, 29], [30, 39], [40, 49], [50, 59], [60, 69], [70, 79], [80, 89], ], tab::SQLTable = person, )

Finds all individuals in age groups as specified by age_groupings.

Arguments:

ids - list of person_id's; each ID must be of subtype Integer

age_groupings - a vector of age groups of the form [[10, 19], [20, 29],] denoting an age group of 10 - 19 and 20 - 29 respectively; age values must subtype of Integer

conn - database connection using DBInterface

Keyword Arguments:

age_groupings - a vector of age groups of the form [[10, 19], [20, 29],] denoting an age group of 10 - 19 and 20 - 29 respectively; age values must subtype of Integer

minuend - the year that a patient's year_of_birth variable is subtracted from; default :now. There are three different options that can be set:

:now - the year as of the day the code is executed given in UTC time

:db - the last year that any record was found in the database using the "observation_period" table (considered by OHDSI experts to have the latest records in a database)

any year provided by a user as long as it is an Integer (such as 2022, 1998, etc.)

tab::SQLTable - the SQLTable representing the Person table; default person

Returns

df::DataFrame - a two column DataFrame comprised of columns: :person_id and :age_group

Note

Age can be difficult to be calculated consistently. In this case, there are some assumptions made to ensure consistency:

According to the OMOP CDM v5.4, only the variable year_of_birth is guaranteed for a given patient. This is one of three options used as the minuend in age calculations.

The subtrahend is based on what one chooses for the minuend key word argument.

The age is then calculated following what is selected based on 1 and 2. This flexibility is encoded to allow a user to choose how they want age groups calculated as well as clear up an ambiguity on how this is determined.

OMOPCDMCohortCreator.GetPatientVisits — Function

GetPatientVisits(ids, conn; tab::SQLTable = visit_occurrence)

Given a list of person IDs, find all their visits.

Arguments:

ids - list of person_id's; each ID must be of subtype Integer conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Visit Occurrence table; default visit_occurrence

Returns

df::DataFrame - a two column DataFrame comprised of columns: :person_id and :visit_occurrence_id

OMOPCDMCohortCreator.GetMostRecentConditions — Function

GetMostRecentConditions(ids, conn; tab::SQLTable = condition_occurrence)

Given a list of person IDs, find their last recorded conditions.

Arguments:

ids - list of person_id's; each ID must be of subtype Integer conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Condition Occurrence table; default condition_occurrence

Returns

df::DataFrame - a two column DataFrame comprised of columns: :person_id and :condition_concept_id

OMOPCDMCohortCreator.GetMostRecentVisit — Function

GetMostRecentVisit(ids, conn; tab::SQLTable = visit_occurrence)

Given a list of person IDs, find their last recorded visit.

Arguments:

ids - list of person_id's; each ID must be of subtype Integer conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Visit Occurrence table; default visit_occurrence

Returns

df::DataFrame - a two column DataFrame comprised of columns: :person_id and :visit_occurrence_id

OMOPCDMCohortCreator.GetVisitCondition — Function

GetVisitCondition(visitids, conn; tab::SQLTable = visitoccurrence)

Given a list of visit IDs, find their corresponding conditions.

Arguments:

visit_ids - list of visit_id's; each ID must be of subtype Integer conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Condition Occurrence table; default condition_occurrence

Returns

df::DataFrame - a two column DataFrame comprised of columns: :visit_occurrence_id and :condition_concept_id

Filters

These functions accepts parameters to produce queries that look for specific subpopulations or information given specific patient identifier(s) (i.e. person_id).

OMOPCDMCohortCreator.VisitFilterPersonIDs — Function

VisitFilterPersonIDs(visitcodes, conn; tab::SQLTable = visitoccurrence)

Given a list of visit concept IDs, visit_codes return from the database patients matching at least one of the provided visit codes from the Visit Occurrence table.

Arguments:

visit_codes - a vector of visit_concept_id's; must be a subtype of Integer

conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Visit Occurrence table; default visit_occurrence

Returns

ids::Vector{Int64} - the list of persons resulting from the filter

OMOPCDMCohortCreator.ConditionFilterPersonIDs — Function

ConditionFilterPersonIDs(conditioncodes, conn; tab::SQLTable = conditionoccurrence)

Given a list of condition concept IDs, condition_codes, return from the database individuals having at least one entry in the Condition Occurrence table matching at least one of the provided condition types.

Arguments:

condition_codes - a vector of condition_concept_id's; must be a subtype of Integer

conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Condition Occurrence table; default condition_occurrence

Returns

ids::Vector{Int64} - the list of persons resulting from the filter

OMOPCDMCohortCreator.RaceFilterPersonIDs — Function

RaceFilterPersonIDs(race_codes, conn; tab::SQLTable = person)

Given a list of condition concept IDs, race_codes, return from the database individuals having at least one entry in the Person table matching at least one of the provided race types.

Arguments:

race_codes - a vector of race_concept_id's; must be a subtype of Integer

conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Person table; default person

Returns

ids::Vector{Int64} - the list of persons resulting from the filter

OMOPCDMCohortCreator.GenderFilterPersonIDs — Function

GenderFilterPersonIDs(gendercodes, conn; tab::SQLTable = visitoccurrence)

Given a list of visit concept IDs, gender_codes return from the database individuals having at least one entry in the Person table matching at least one of the provided gender types.

Arguments:

visit_codes - a vector of gender_concept_id's; must be a subtype of Integer

conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Person table; default person

Returns

ids::Vector{Int64} - the list of persons resulting from the filter

OMOPCDMCohortCreator.StateFilterPersonIDs — Function

StateFilterPersonIDs(states, conn; tab::SQLTable = location, join_tab::SQLTable = person)

Given a list of states, states, return from the database individuals found in the provided state list.

Arguments:

states - a vector of state abbreviations; must be a subtype of AbstractString

conn - database connection using DBInterface

Keyword Arguments:

tab::SQLTable - the SQLTable representing the Location table; default location

join_tab::SQLTable - the SQLTable representing the Person table; default person

Returns

ids::Vector{Int64} - the list of persons resulting from the filter

Missing docstring.

Missing docstring for AgeGroupFilterPersonIDs. Check Documenter's build log for details.

Generators

The generator functions are to set generate initial connections to an OMOP CDM database or to finalize resulting data from queries into various outputs.

OMOPCDMCohortCreator.GenerateDatabaseDetails — Function

GenerateDatabaseDetails(dialect::Symbol, schema::String)

Generates the dialect and schema details for accessing a given OMOP CDM database.

Arguments:

dialect::Symbol - the dialect used for SQL queries (to see what is dialects are available, see here: https://mechanicalrabbit.github.io/FunSQL.jl/stable/reference/#FunSQL.SQLDialect)

schema::String - the name of the database schema being used.

OMOPCDMCohortCreator.GenerateGroupCounts — Function

GenerateGroupCounts(data::DataFrame)

Given data in a DataFrame, get group counts based on each feature found in the DataFrame and removes person_id for privacy aggregation purposes.

Arguments:

data::DataFrame - a DataFrame that must have at least a person_id column

Returns:

df::DataFrame - a DataFrame that contains the group counts based on each feature found in data with the person_id field removed for privacy

OMOPCDMCohortCreator.GenerateTables — Function

GenerateTables(conn; inplace = true, exported = false)

Generates Julia representations of all tables found in a given OMOP CDM database.

Arguments:

dialect::Symbol - the dialect used for SQL queries (to see what is dialects are available, see here: https://mechanicalrabbit.github.io/FunSQL.jl/stable/reference/#FunSQL.SQLDialect)

schema::String - the name of the database schema being used.

Executors

These functions perform quality assurance checks on data extracts genereated from OMOPCDMCohortCreator queries.

OMOPCDMCohortCreator.ExecuteAudit — Function

TODO: Add docs when ready